Update testing on sieve

+++ other/partition_values.cpp
@@ -18,7 +18,7 @@ int main() {
     //uint64_t max = 1048576; //2^20
     auto visited = boost::dynamic_bitset<>(max + 1);
 
-    int thresh = 10;
+    uint64_t thresh = 10;
 
     // Wipe the files
     ofstream alg2_file;
@@ -30,8 +30,8 @@ int main() {
 
     HighestPrimeSieve s = HighestPrimeSieve(max);
 
-    int p_thresh = 1;
-    for (int i = 0; i < thresh; i++) {
+    uint64_t p_thresh = 1;
+    for (uint64_t i = 0; i < thresh; i++) {
         p_thresh = s.next_prime(p_thresh);
     }
     cout << p_thresh << endl;
